quite important changes in matrix and vector classes
so that they are really serializable. further, similar arrangement will follow in network class as well, because that too, as it is now, not serializable. although al tests are green, i am not so sure (1) if matrix construction is entirely right (2) if it is optimal... think should be made much simpler